Long to Integer

+++ src/main/java/ua/com/javarush/gnew/contactm/repository/ContactRepository.java
@@ -3,6 +3,6 @@ package ua.com.javarush.gnew.contactm.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.com.javarush.gnew.contactm.entity.Contact;
 
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
 
 }
